requirements.txt
tqdm
numpy
scikit-image
scipy
matplotlib
opencv-python
pillow